requirements.txt
# Telegram Bot для маркетинговой аналитики "Евгенич СПБ"
python-telegram-bot==20.7
gspread==5.12.0
pandas==2.1.4
requests==2.31.0
APScheduler==3.10.4
redis==5.0.1
matplotlib==3.8.2
seaborn==0.13.0
aiohttp==3.9.1
python-dotenv==1.0.0
numpy==1.26.2
openpyxl==3.1.2
Pillow==10.1.0
google-auth==2.25.2
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
asyncpg==0.29.0
